All 512 file setups in one spot
I added the code required for each of the files inline with the instructions.  You said to make a new page, but I couldn't quite figure that out.  I tried to make it VERY clear that the user needs to change these settings to match their pump.  Hopefully this helps! :)